requirements.txt
attrs==20.3.0
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
coverage==5.3
docopt==0.6.2
idna==2.10
importlib-metadata==3.1.0
iniconfig==1.1.1
joblib==0.17.0
ml2en==1.0.0
nltk==3.5
numpy==1.19.4
packaging==20.7
pandas==1.1.4
pipreqs==0.4.10
pkg-resources==0.0.0
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pyspellchecker==0.5.5
pytest==6.1.2
pytest-cov==2.10.1
python-dateutil==2.8.1
pytz==2020.4
regex==2020.11.13
requests==2.25.0
scikit-learn==0.23.2
scipy==1.5.4
six==1.15.0
sklearn==0.0
threadpoolctl==2.1.0
toml==0.10.2
tqdm==4.54.0
urllib3==1.26.2
yarg==0.1.9
zipp==3.4.0